moved utils functions from graph to graph utils, removed unnecessary extension functions
