+++ src/main/java/at/hannibal2/skyhanni/utils/GraphUtils.kt
@@ -1,10 +1,10 @@
 package at.hannibal2.skyhanni.utils
 
+import at.hannibal2.skyhanni.data.model.DijkstraTree
 import at.hannibal2.skyhanni.data.model.Graph
 import at.hannibal2.skyhanni.data.model.GraphNode
-import at.hannibal2.skyhanni.data.model.findAllShortestDistances
-import at.hannibal2.skyhanni.data.model.findDijkstraDistances
 import at.hannibal2.skyhanni.data.model.findPathToDestination
+import java.util.PriorityQueue
 import java.util.Stack
 
 object GraphUtils {
@@ -12,11 +12,10 @@ object GraphUtils {
      * Find the fastest path from [closestNode] to *any* node that matches [condition].
      */
     fun findFastestPath(
-        graph: Graph,
         closestNode: GraphNode,
         condition: (GraphNode) -> Boolean,
     ): Pair<Graph, Double>? {
-        val distances = graph.findDijkstraDistances(closestNode, condition)
+        val distances = findDijkstraDistances(closestNode, condition)
         val entry = distances.lastVisitedNode.takeIf(condition)
         return entry?.let {
             distances.findPathToDestination(it)
@@ -34,7 +33,7 @@ object GraphUtils {
         val paths = mutableMapOf<GraphNode, Graph>()
 
         val map = mutableMapOf<GraphNode, Double>()
-        val distances = graph.findAllShortestDistances(closestNode)
+        val distances = findAllShortestDistances(closestNode)
         for (graphNode in graph.nodes) {
             if (!condition(graphNode)) continue
             val (path, distance) = distances.findPathToDestination(graphNode)
@@ -65,4 +64,66 @@ object GraphUtils {
         }
         return allClusters
     }
+
+    fun findShortestPathAsGraph(start: GraphNode, end: GraphNode): Graph = findShortestPathAsGraphWithDistance(start, end).first
+
+    /**
+     * Find a tree of distances to the [start] node using dijkstra's algorithm.
+     */
+    fun findDijkstraDistances(
+        start: GraphNode,
+        /**
+         * Bail out early before collecting all the distances to all nodes in the graph. This will not collect valid distance data for *all*
+         * nodes for which bailout matches, but only the closest one.
+         */
+        bailout: (GraphNode) -> Boolean,
+    ): DijkstraTree {
+        val distances = mutableMapOf<GraphNode, Double>()
+        val previous = mutableMapOf<GraphNode, GraphNode>()
+        val visited = mutableSetOf<GraphNode>()
+        val queue = PriorityQueue<GraphNode>(compareBy { distances.getOrDefault(it, Double.MAX_VALUE) })
+        var lastVisitedNode: GraphNode = start
+
+        distances[start] = 0.0
+        queue.add(start)
+
+        while (queue.isNotEmpty()) {
+            val current = queue.poll()
+            lastVisitedNode = current
+            if (bailout(current)) break
+
+            visited.add(current)
+
+            current.neighbours.forEach { (neighbour, weight) ->
+                if (neighbour !in visited) {
+                    val newDistance = distances.getValue(current) + weight
+                    if (newDistance < distances.getOrDefault(neighbour, Double.MAX_VALUE)) {
+                        distances[neighbour] = newDistance
+                        previous[neighbour] = current
+                        queue.add(neighbour)
+                    }
+                }
+            }
+        }
+
+        return DijkstraTree(
+            start,
+            distances,
+            previous,
+            lastVisitedNode,
+        )
+    }
+
+    fun findAllShortestDistances(start: GraphNode): DijkstraTree {
+        return findDijkstraDistances(start) { false }
+    }
+
+    fun findShortestPathAsGraphWithDistance(start: GraphNode, end: GraphNode): Pair<Graph, Double> {
+        val distances = findDijkstraDistances(start) { it == end }
+        return distances.findPathToDestination(end)
+    }
+
+    fun findShortestPath(start: GraphNode, end: GraphNode): List<LorenzVec> = findShortestPathAsGraph(start, end).toPositionsList()
+
+    fun findShortestDistance(start: GraphNode, end: GraphNode): Double = findShortestPathAsGraphWithDistance(start, end).second
 }
